moving more code into ipysnobal, deleted initialize_model.py and smrf_ipysnobal.py
