+++ awsm/interface/__init__.py
@@ -3,5 +3,5 @@
 __author__ = """Micah Sandusky"""
 __email__ = 'micah.sandusky@ars.usda.gov'
 
-from . import interface, ipysnobal, smrf_ipysnobal, initialize_model, \
+from . import interface, ipysnobal, smrf_ipysnobal, \
     pysnobal_io, ingest_data
